# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/abonnee/edit/{id}": {
            "put": {
                "tags": [
                    "Abonnee"
                ],
                "summary": "Update a contrat",
                "operationId": "93d1c2c1c305dd5be10d0e7824c25667",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the contrat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "statut",
                                    "slug",
                                    "utilisateur_id",
                                    "typeabonnement_id"
                                ],
                                "properties": {
                                    "statut": {
                                        "type": "string",
                                        "example": "En cours"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "slug"
                                    },
                                    "utilisateur_id": {
                                        "type": "string",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "abonnee updated"
                    },
                    "404": {
                        "description": "abonnee not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/abonnee/create": {
            "post": {
                "tags": [
                    "Abonnee"
                ],
                "summary": "Create a new user",
                "operationId": "6ec003cce64bdb57fc00fe1dac085810",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "statut",
                                    "slug",
                                    "utilisateur_id",
                                    "typeabonnement_id"
                                ],
                                "properties": {
                                    "statut": {
                                        "type": "string",
                                        "example": "En cours"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "slug"
                                    },
                                    "utilisateur_id": {
                                        "type": "string",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Abonnee created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/abonnee/show/{id}": {
            "get": {
                "tags": [
                    "Abonnee"
                ],
                "summary": "Get a specific resource",
                "operationId": "cc3617c9e76de7aef8897d3e9764d42e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": " id de abonnee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
        "/api/v1/abonnee/{id}": {
            "delete": {
                "tags": [
                    "Abonnee"
                ],
                "summary": "Delete a user",
                "operationId": "248ba213011e2746015c686414952e65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    }
}